extract the `AsBindGroup` logic to a separate module instead of a struct
