+++ crates/bevy_pbr/src/extended_material.rs
@@ -4,10 +4,11 @@ use bevy_material::{AlphaMode, OpaqueRendererMethod};
 use bevy_mesh::MeshVertexBufferLayoutRef;
 use bevy_reflect::{impl_type_path, Reflect};
 use bevy_render::{
+    combined_bind_group as cbg,
     render_resource::{
         AsBindGroup, AsBindGroupError, BindGroupBuilder, BindGroupLayout, BindGroupLayoutEntry,
-        BindlessDescriptor, BindlessSlabResourceLimit, CombinedBindGroup as Cbg,
-        RenderPipelineDescriptor, SpecializedMeshPipelineError,
+        BindlessDescriptor, BindlessSlabResourceLimit, RenderPipelineDescriptor,
+        SpecializedMeshPipelineError,
     },
     renderer::RenderDevice,
 };
@@ -170,27 +171,23 @@ where
 impl_type_path!((in bevy_pbr::extended_material) ExtendedMaterial<B: Material, E: MaterialExtension>);
 
 impl<B: Material, E: MaterialExtension> AsBindGroup for ExtendedMaterial<B, E> {
-    type Data = <Cbg<'static, B, E> as AsBindGroup>::Data;
-    type Param = <Cbg<'static, B, E> as AsBindGroup>::Param;
+    type Data = cbg::CombinedBindGroupData<B::Data, E::Data>;
+    type Param = (B::Param, E::Param);
 
     fn bindless_slot_count() -> Option<BindlessSlabResourceLimit> {
-        Cbg::<'static, B, E>::bindless_slot_count()
+        cbg::bindless_slot_count::<B, E>()
     }
 
     fn bindless_supported(render_device: &RenderDevice) -> bool {
-        Cbg::<'static, B, E>::bindless_supported(render_device)
+        B::bindless_supported(render_device) && E::bindless_supported(render_device)
     }
 
     fn label() -> &'static str {
-        Cbg::<'static, B, E>::label()
+        E::label()
     }
 
     fn bind_group_data(&self) -> Self::Data {
-        Cbg {
-            base: &self.base,
-            extension: &self.extension,
-        }
-        .bind_group_data()
+        cbg::bind_group_data(&self.base, &self.extension)
     }
 
     fn build_bind_group(
@@ -198,14 +195,18 @@ impl<B: Material, E: MaterialExtension> AsBindGroup for ExtendedMaterial<B, E> {
         layout: &BindGroupLayout,
         render_device: &RenderDevice,
         param: &mut SystemParamItem<'_, '_, Self::Param>,
-        mut force_no_bindless: bool,
+        force_no_bindless: bool,
         output: &mut BindGroupBuilder,
     ) -> Result<(), AsBindGroupError> {
-        Cbg {
-            base: &self.base,
-            extension: &self.extension,
-        }
-        .build_bind_group(layout, render_device, param, force_no_bindless, output)
+        cbg::build_bind_group(
+            &self.base,
+            &self.extension,
+            layout,
+            render_device,
+            param,
+            force_no_bindless,
+            output,
+        )
     }
 
     fn bind_group_layout_entries(
@@ -215,11 +216,11 @@ impl<B: Material, E: MaterialExtension> AsBindGroup for ExtendedMaterial<B, E> {
     where
         Self: Sized,
     {
-        Cbg::<'static, B, E>::bind_group_layout_entries(render_device, force_no_bindless)
+        cbg::bind_group_layout_entries::<B, E>(render_device, force_no_bindless)
     }
 
     fn bindless_descriptor() -> Option<BindlessDescriptor> {
-        Cbg::<'static, B, E>::bindless_descriptor()
+        cbg::bindless_descriptor::<B, E>()
     }
 }
