+++ crates/bevy_render/src/render_resource/mod.rs
@@ -6,7 +6,6 @@ mod bind_group_layout;
 mod bindless;
 mod buffer;
 mod buffer_vec;
-mod combined_bind_group;
 mod gpu_array_buffer;
 mod pipeline;
 mod pipeline_cache;
@@ -24,7 +23,6 @@ pub use bind_group_layout::*;
 pub use bindless::*;
 pub use buffer::*;
 pub use buffer_vec::*;
-pub use combined_bind_group::*;
 pub use gpu_array_buffer::*;
 pub use pipeline::*;
 pub use pipeline_cache::*;
